Consolidate PlantUML service
- Use the same service for both `/c4plantuml` and `/plantuml`
- Upgrade C4 definition files from https://github.com/plantuml-stdlib/C4-PlantUML
- Add c4_dynamic and c4_deployment definition files
- Add tests
