+++ server/src/main/resources/c4_component.puml
@@ -7,46 +7,113 @@
 ' Colors
 ' ##################################
 
-!define COMPONENT_BG_COLOR #85BBF0
+!global $COMPONENT_FONT_COLOR = "#000000"
+!global $COMPONENT_BG_COLOR = "#85BBF0"
+!global $COMPONENT_BORDER_COLOR = "#78A8D8"
+!global $EXTERNAL_COMPONENT_BG_COLOR = "#CCCCCC"
+!global $EXTERNAL_COMPONENT_BORDER_COLOR = "#BFBFBF"
 
 ' Styling
 ' ##################################
 
 skinparam rectangle<<component>> {
-    StereotypeFontColor ELEMENT_FONT_COLOR
-    FontColor #000000
-    BackgroundColor COMPONENT_BG_COLOR
-    BorderColor #78A8D8
+    StereotypeFontColor $COMPONENT_FONT_COLOR
+    FontColor $COMPONENT_FONT_COLOR
+    BackgroundColor $COMPONENT_BG_COLOR
+    BorderColor $COMPONENT_BORDER_COLOR
 }
 
 skinparam database<<component>> {
-    StereotypeFontColor ELEMENT_FONT_COLOR
-    FontColor #000000
-    BackgroundColor COMPONENT_BG_COLOR
-    BorderColor #78A8D8
+    StereotypeFontColor $COMPONENT_FONT_COLOR
+    FontColor $COMPONENT_FONT_COLOR
+    BackgroundColor $COMPONENT_BG_COLOR
+    BorderColor $COMPONENT_BORDER_COLOR
+}
+
+skinparam queue<<component>> {
+    StereotypeFontColor $COMPONENT_FONT_COLOR
+    FontColor $COMPONENT_FONT_COLOR
+    BackgroundColor $COMPONENT_BG_COLOR
+    BorderColor $COMPONENT_BORDER_COLOR
+}
+
+skinparam rectangle<<external_component>> {
+    StereotypeFontColor $COMPONENT_FONT_COLOR
+    FontColor $COMPONENT_FONT_COLOR
+    BackgroundColor $EXTERNAL_COMPONENT_BG_COLOR
+    BorderColor $EXTERNAL_COMPONENT_BORDER_COLOR
+}
+
+skinparam database<<external_component>> {
+    StereotypeFontColor $COMPONENT_FONT_COLOR
+    FontColor $COMPONENT_FONT_COLOR
+    BackgroundColor $EXTERNAL_COMPONENT_BG_COLOR
+    BorderColor $EXTERNAL_COMPONENT_BORDER_COLOR
+}
+
+skinparam queue<<external_component>> {
+    StereotypeFontColor $COMPONENT_FONT_COLOR
+    FontColor $COMPONENT_FONT_COLOR
+    BackgroundColor $EXTERNAL_COMPONENT_BG_COLOR
+    BorderColor $EXTERNAL_COMPONENT_BORDER_COLOR
 }
 
 ' Layout
 ' ##################################
 
-!definelong LAYOUT_WITH_LEGEND
+!procedure LAYOUT_WITH_LEGEND()
 hide stereotype
 legend right
-|=              |= Type |
-|<PERSON_BG_COLOR>      | person |
-|<EXTERNAL_PERSON_BG_COLOR>      | external person |
-|<SYSTEM_BG_COLOR>   | system |
-|<EXTERNAL_SYSTEM_BG_COLOR>      | external system |
-|<CONTAINER_BG_COLOR>   | container |
-|<COMPONENT_BG_COLOR>   | component |
+|<color:$LEGEND_TITLE_COLOR>**Legend**</color> |
+|<$PERSON_BG_COLOR> person |
+|<$SYSTEM_BG_COLOR> system |
+|<$CONTAINER_BG_COLOR> container |
+|<$COMPONENT_BG_COLOR> component |
+|<$EXTERNAL_PERSON_BG_COLOR> external person |
+|<$EXTERNAL_SYSTEM_BG_COLOR> external system |
+|<$EXTERNAL_CONTAINER_BG_COLOR> external container |
+|<$EXTERNAL_COMPONENT_BG_COLOR> external component |
 endlegend
-!enddefinelong
+!endprocedure
 
 ' Elements
 ' ##################################
 
-!define Component(e_alias, e_label, e_techn) rectangle "==e_label\n//<size:TECHN_FONT_SIZE>[e_techn]</size>//" <<component>> as e_alias
-!define Component(e_alias, e_label, e_techn, e_descr) rectangle "==e_label\n//<size:TECHN_FONT_SIZE>[e_techn]</size>//\n\n e_descr" <<component>> as e_alias
+!function $getComponent($label, $techn, $descr, $sprite)
+!if ($descr == "") && ($sprite == "")
+!return '=='+$label+'\n//<size:'+$TECHN_FONT_SIZE+'>['+$techn+']</size>//'
+!endif
+!if ($descr == "") && ($sprite != "")
+!return '<$'+$sprite+'>\n=='+$label+'\n//<size:'+$TECHN_FONT_SIZE+'>['+$techn+']</size>//'
+!endif
+!if ($descr != "") && ($sprite == "")
+!return '=='+$label+'\n//<size:'+$TECHN_FONT_SIZE+'>['+$techn+']</size>//\n\n '+$descr
+!endif
+!if ($descr != "") && ($sprite != "")
+!return '<$'+$sprite+'>\n=='+$label+'\n//<size:'+$TECHN_FONT_SIZE+'>['+$techn+']</size>//\n\n '+$descr
+!endif
+!endfunction
 
-!define ComponentDb(e_alias, e_label, e_techn) database "==e_label\n//<size:TECHN_FONT_SIZE>[e_techn]</size>//" <<component>> as e_alias
-!define ComponentDb(e_alias, e_label, e_techn, e_descr) database "==e_label\n//<size:TECHN_FONT_SIZE>[e_techn]</size>//\n\n e_descr" <<component>> as e_alias
+!unquoted procedure Component($alias, $label, $techn, $descr="", $sprite="")
+rectangle "$getComponent($label, $techn, $descr, $sprite)" <<component>> as $alias
+!endprocedure
+
+!unquoted procedure ComponentDb($alias, $label, $techn, $descr="", $sprite="")
+database "$getComponent($label, $techn, $descr, $sprite)" <<component>> as $alias
+!endprocedure
+
+!unquoted procedure ComponentQueue($alias, $label, $techn, $descr="", $sprite="")
+queue "$getComponent($label, $techn, $descr, $sprite)" <<component>> as $alias
+!endprocedure
+
+!unquoted procedure Component_Ext($alias, $label, $techn, $descr="", $sprite="")
+rectangle "$getComponent($label, $techn, $descr, $sprite)" <<external_component>> as $alias
+!endprocedure
+
+!unquoted procedure ComponentDb_Ext($alias, $label, $techn, $descr="", $sprite="")
+database "$getComponent($label, $techn, $descr, $sprite)" <<external_component>> as $alias
+!endprocedure
+
+!unquoted procedure ComponentQueue_Ext($alias, $label, $techn, $descr="", $sprite="")
+queue "$getComponent($label, $techn, $descr, $sprite)" <<external_component>> as $alias
+!endprocedure
